move BaseService to model package
